Ensure that integer division produces float in Py2.
This bug was filed under [issue 2](https://github.com/mgiraldo/tenprintcover-py/issues/2) and was caused by the different semantics of the integer division operator `/` on Python 2 and Python 3. See also [PEP 238](https://www.python.org/dev/peps/pep-0238/). Because the code expects that integer division produces a float (as it does on Python 3) this change imports the correct division operator on Python 2.